pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend="hatchling.build"

[project]
name = "validation-dags"
authors = [{name = "the airflow-laminar authors", email = "t.paine154@gmail.com"}]
description = "DAGs to validate airflow-laminar packages"
readme = "README.md"
license = { text = "Apache-2.0" }
version = "0.1.0"
requires-python = ">=3.11"
keywords = [
    "airflow",
    "scheduler",
    "dag",
    "dags",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "apache-airflow>=2.10.3",
    "apache-airflow-providers-ssh",
    "airflow-balancer",
    "airflow-common-operators",
    "airflow-config",
    "airflow-supervisor",
]

[project.optional-dependencies]
develop = [
    "build",
    "bump-my-version",
    "check-manifest",
    "hatchling",
    "pytest",
    "pytest-cov",
    "ruff",
    "twine",
    "uv",
    "wheel",
]

[project.scripts]

[project.urls]
Repository = "https://github.com/airflow-laminar/validation-dags"
Homepage = "https://github.com/airflow-laminar/validation-dags"

[tool.bumpversion]
current_version = "0.1.0"
commit = true
tag = true

[[tool.bumpversion.files]]
filename = "validation_dags/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[tool.check-manifest]
ignore = [
    ".copier-answers.yaml",
    "Makefile",
    "docs/**/*",
]

[tool.coverage.run]
branch = true
omit = [
    "validation_dags/tests/integration/",
]
[tool.coverage.report]
exclude_also = [
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@(abc\\.)?abstractmethod",
]
ignore_errors = true
fail_under = 50

[tool.hatch.build]
artifacts = []

[tool.hatch.build.sources]
src = "/"

[tool.hatch.build.targets.sdist]
packages = ["validation_dags"]

[tool.hatch.build.targets.wheel]
packages = ["validation_dags"]

[tool.hatch.build.targets.wheel.shared-data]

[tool.pytest.ini_options]
addopts = ["-vvv", "--junitxml=junit.xml"]
testpaths = "validation_dags/tests"

[tool.ruff]
line-length = 150

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort]
combine-as-imports = true
default-section = "third-party"
known-first-party = ["validation_dags"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
